requirements.txt
opencv-python
numpy
tensorflow
pyttsx3
Pillow
mediapipe
scikit-learn 